+++ Projects/Animal Trainer/Sprites.cs
@@ -1,27 +1,16 @@
-﻿using System.Linq;
-
-namespace Animal_Trainer;
+﻿namespace Animal_Trainer;
 
 public static class Sprites
 {
+	public const int Width = 7;
+	public const int Height = 5;
+
 	public const string Open =
 		@"       " + "\n" +
 		@"       " + "\n" +
 		@"       " + "\n" +
 		@"       " + "\n" +
 		@"       ";
-	public const string Town =
-		@"╔══ ══╗" + "\n" +
-		@"║▀▀ ▀▀║" + "\n" +
-		@"║▀▀ ▀▀║" + "\n" +
-		@"║▄▄ ▄▄║" + "\n" +
-		@"╚══ ══╝";
-	public const string Castle =
-		@"╔══ ══╗" + "\n" +
-		@"║╔═══╗║" + "\n" +
-		@"║║   ║║" + "\n" +
-		@"║╚═══╝║" + "\n" +
-		@"╚══ ══╝";
 	public const string BuildingSmall =
 		@"       " + "\n" +
 		@" /---\ " + "\n" +
@@ -221,9 +210,9 @@ public static class Sprites
 		@"▀▄▀▄▀▄▀" + "\n" +
 		@"▀▄▀▄▀▄▀";
 	public const string Tree =
-		@" /---\ " + "\n" +
-		@" \___/ " + "\n" +
-		@"  | |  " + "\n" +
+		@" /‾‾‾\ " + "\n" +
+		@"/‾\ /‾\" + "\n" +
+		@"\_____/" + "\n" +
 		@"  | |  " + "\n" +
 		@"  / \  ";
 	public const string Tree2 =
@@ -517,153 +506,4 @@ public static class Sprites
 
 	public static readonly string[] IdlePlayer =
 		Enumerable.Repeat(Idle1, 100).Concat(Enumerable.Repeat(Idle2, 10)).ToArray();
-
-	// would be nice to give up/down their own animations, but
-	// for now they are just copies of left/right
-	public static readonly string[] Idle = (string[])IdlePlayer.Clone();
-
-
-	public static readonly string[] IdleBoar = new[]
-	{
-		// 0
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 1
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 2
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ', ,' ",
-		// 3
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ', ,' ",
-		// 4
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ', ,' ",
-		// 5
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ', ,' ",
-		// 6
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ', ,' ",
-		// 7
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,' ', ",
-		// 8
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,' ', ",
-		// 9
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,' ', ",
-		// 10
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,' ', ",
-		// 11
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,' ', ",
-		// 12
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 13
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 14
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 15
-		@" ^..^  " + '\n' +
-		@"(oo)\  " + '\n' +
-		@"  O  )~" + '\n' +
-		@" '' ,, ",
-		// 16
-		@" ^..^  " + '\n' +
-		@"(oo)\  " + '\n' +
-		@"  O  )~" + '\n' +
-		@" '' ,, ",
-		// 17
-		@" ^..^  " + '\n' +
-		@"(oo)\  " + '\n' +
-		@"  O  )~" + '\n' +
-		@" '' ,, ",
-		// 18
-		@" ^..^  " + '\n' +
-		@"(oo)\  " + '\n' +
-		@"  O  )~" + '\n' +
-		@" '' ,, ",
-		// 19
-		@" ^..^  " + '\n' +
-		@"(oo)\  " + '\n' +
-		@"  O  )~" + '\n' +
-		@" '' ,, ",
-		// 20
-		@"       " + '\n' +
-		@"^--^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 21
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 22
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 23
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 24
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 25
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 26
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-		// 27
-		@"       " + '\n' +
-		@"^..^   " + '\n' +
-		@"(oo) )~" + '\n' +
-		@" ,, ,, ",
-	};
 }
